requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.0
langchain==0.1.4
langchain-groq==0.0.1
selenium==4.17.2
beautifulsoup4==4.12.2
lxml==5.1.0
